# components/employee-mgt-service/openapi.yaml
openapi: 3.0.3
info:
  title: Employee Management API
  description: |
    A simple REST API for managing employee records.
    This service provides endpoints to create, retrieve, and list employees.
    Employee data is stored in memory and will be reset when the service restarts.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Employees
    description: Operations related to employee management

paths:
  /employee-mgt/employees:
    get:
      tags:
        - Employees
      summary: Get all employees
      description: |
        Retrieves a list of all employees stored in the system.
        Returns an empty array if no employees exist.
      operationId: getAllEmployees
      responses:
        '200':
          description: Successfully retrieved list of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
              examples:
                employees:
                  summary: List of employees
                  value:
                    - id: 1
                      firstName: John
                      lastName: Doe
                      email: john.doe@example.com
                      department: Engineering
                      position: Software Engineer
                    - id: 2
                      firstName: Jane
                      lastName: Smith
                      email: jane.smith@example.com
                      department: Marketing
                      position: Marketing Manager

    post:
      tags:
        - Employees
      summary: Add a new employee
      description: |
        Creates a new employee record in the system.
        The employee ID is auto-generated by the server.
        FirstName, LastName, and Email are required fields.
      operationId: addEmployee
      requestBody:
        description: Employee object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeInput'
            examples:
              newEmployee:
                summary: New employee request
                value:
                  firstName: John
                  lastName: Doe
                  email: john.doe@example.com
                  department: Engineering
                  position: Software Engineer
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
              examples:
                createdEmployee:
                  summary: Created employee response
                  value:
                    id: 1
                    firstName: John
                    lastName: Doe
                    email: john.doe@example.com
                    department: Engineering
                    position: Software Engineer
        '400':
          description: Invalid request body or missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation error
                  value:
                    message: "FirstName, LastName, and Email are required"
                    code: 400

  /employee-mgt/employees/{employeeId}:
    get:
      tags:
        - Employees
      summary: Get employee by ID
      description: |
        Retrieves a specific employee by their unique identifier.
        Returns a 404 error if the employee does not exist.
      operationId: getEmployeeById
      parameters:
        - name: employeeId
          in: path
          description: Unique identifier of the employee
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      responses:
        '200':
          description: Successfully retrieved employee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
              examples:
                employee:
                  summary: Employee details
                  value:
                    id: 1
                    firstName: John
                    lastName: Doe
                    email: john.doe@example.com
                    department: Engineering
                    position: Software Engineer
        '400':
          description: Invalid employee ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidId:
                  summary: Invalid ID error
                  value:
                    message: "Invalid employee ID"
                    code: 400
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: Not found error
                  value:
                    message: "Employee not found"
                    code: 404

components:
  schemas:
    Employee:
      type: object
      description: Represents an employee record with all fields including the auto-generated ID
      required:
        - id
        - firstName
        - lastName
        - email
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the employee (auto-generated)
          example: 1
          readOnly: true
        firstName:
          type: string
          description: First name of the employee
          example: John
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          description: Last name of the employee
          example: Doe
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          description: Email address of the employee
          example: john.doe@example.com
        department:
          type: string
          description: Department where the employee works
          example: Engineering
          maxLength: 100
        position:
          type: string
          description: Job position or title of the employee
          example: Software Engineer
          maxLength: 100

    EmployeeInput:
      type: object
      description: Input schema for creating a new employee (ID is auto-generated)
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
          description: First name of the employee
          example: John
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          description: Last name of the employee
          example: Doe
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          description: Email address of the employee
          example: john.doe@example.com
        department:
          type: string
          description: Department where the employee works
          example: Engineering
          maxLength: 100
        position:
          type: string
          description: Job position or title of the employee
          example: Software Engineer
          maxLength: 100

    ErrorResponse:
      type: object
      description: Standard error response format
      required:
        - message
        - code
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Employee not found"
        code:
          type: integer
          description: HTTP status code
          example: 404